refactored seeds file

+++ db/seeds.rb
@@ -4,15 +4,18 @@ NUM_PRESENTATIONS = 8
 PARTICIPATIONS_PER_PRESENTATION = 10
 SURVEYS_PER_PRESENTATION = 3
 QUESTIONS_PER_SURVEY = 5
+RESPONSE_NUM_MAX = 5
 PASSWORD = "testing"
 
-puts "Destroying everything..."
-User.destroy_all
-Presentation.destroy_all
-Participation.destroy_all
-Survey.destroy_all
-Question.destroy_all
-Response.destroy_all
+unless Rails.env.prod?
+  puts "Destroying everything..."
+  User.destroy_all
+  Presentation.destroy_all
+  Participation.destroy_all
+  Survey.destroy_all
+  Question.destroy_all
+  Response.destroy_all
+end
 
 puts "Creating basic users..."
 NUM_USERS.times do |n|
@@ -26,15 +29,16 @@ NUM_USERS.times do |n|
   u.save
 end
 
-puts "Setting #{NUM_ADMINS} users as admins..."
-admin_count = 0
-while admin_count != NUM_ADMINS
-  u = User.all.sample
-  unless u.is_admin
-    u.is_admin = true
-    u.save
-    admin_count += 1
-  end
+puts "Creating #{NUM_ADMINS} admins..."
+NUM_ADMINS.times do |n|
+  u = User.new(
+    first_name: "Admin",
+    last_name: "#{n}",
+    password: PASSWORD,
+    is_admin: true
+  )
+  u.email = "admin#{n}@example.com"
+  u.save
 end
 
 puts "Creating non-admin Khoi..."
@@ -72,10 +76,9 @@ NUM_PRESENTATIONS.times do
     QUESTIONS_PER_SURVEY.times do |ques_num|
       ques = survey.questions.create(
         order: ques_num,
-        prompt: Faker::Hipster.sentence,
-        response_type: rand(2) == 0 ? 'text' : 'number'
+        prompt: Faker::Hipster.words(5).join(' ') + '?',
+        response_type: ['text', 'number'].sample
       )
-      ques.prompt[-1] = '?'
       ques.save
     end
   end
@@ -109,7 +112,7 @@ Question.all.each do |ques|
     Response.create(
       question_id: ques.id,
       user_id: user.id,
-      value: ques.response_type == 'text' ? Faker::Hipster.words(2).join(' ') : rand(1..5).to_s
+      value: ques.response_type == 'text' ? Faker::Hipster.words(2).join(' ') : rand(1..RESPONSE_NUM_MAX).to_s
     )
   end
 end
